+++ src/main/java/com/example/booksserver/userstate/UserBaseFilters.java
@@ -3,10 +3,11 @@ package com.example.booksserver.userstate;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @AllArgsConstructor
 @Getter
 public class UserBaseFilters {
-    private final List<Double> pricesRange;
+    private final List<BigDecimal> pricesRange;
 }
